Sort low win rate tanks last with more battles.
These tanks require more work to improve, so they should be lower
in the least as the intention of the sort is help decide which
tanks are most worth playing.

High battle count is good for high win rate tanks since you can
accomplish two goals at once (improving the tank, and the overall
period).

Also, convert to using a logger instance.